+++ src/main/java/com/menu/menu/mapper/MenuStepMapper.java
@@ -6,6 +6,6 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface MenuStepMapper extends BaseMapper<MenuStep> {
-    List<MenuStep> selectByMenuId(@Param("menuId") Long menuId);
+    List<MenuStep> selectByMenuId(@Param("menuId") Integer menuId);
     int batchInsert(@Param("steps") List<MenuStep> steps);
 }
